2回目のレビュー修正
それぞれ修正しました。
