+++ chapter1/Question1-1-2/MyClass.cs
@@ -1,0 +1,13 @@
+﻿namespace Question1_1_2 {
+    //MyClassとMyStruct
+    //クラス
+    class MyClass {
+        public int X { get; set; }
+        public int Y { get; set; }
+
+        public MyClass(int x, int y) {
+            this.X = x;
+            this.Y = y;
+        }
+    }
+}
