+++ chapter1/Question1-1-2/MyStruct.cs
@@ -1,0 +1,12 @@
+﻿namespace Question1_1_2 {
+    //構造体
+    struct MyStruct {
+        public int X { get; set; }
+        public int Y { get; set; }
+
+        public MyStruct(int x, int y) {
+            this.X = x;
+            this.Y = y;
+        }
+    }
+}
